Path prefix for bucket key #4
Add new option -p to set a prefix file path which will cause uploaded files to
go to a subfolder of the S3 bucket, allowing for multiple git repos to be
hosted in a single bucket.

This works well for new repos but will currently have some pretty unexpected
handling if you've already used git-s3-push. Will need to either ignore the ref
when changing prefix or create a per-prefix ref.